feat: download xrefs as needed and support devsite:// xrefs (#72)
This changes xref support to require libraries to specify the exact xref
files we need. This has two clear benefits:

* Efficiency. We're no longer downloading every xrefmap for every
  library.
* Dependency versioning support. If there are multiple versions of the
  same package (1.0.0, 2.0.0, etc.), they would all register a similar set
  of xrefmaps. If we download and plug in all of those, the behavior would
  be undefined and probably wrong.

I chose to log an error, but not fail the build, when an xref fails to resolve.
That way, if docs are generated in the "wrong" order, we can still make
progress.